feat: Disable progress bar logging at inference time.
Add a progress bar flag in the CorefModel constructor.
Add a default value in the Spacy configuration.
Add the constructor argument with default value set to True in subclasses.
Move the core of _inference to a separate method to avoid duplicate code.
Display the progress bar in _inference only if the progress bar flag is set.